refactor(dashboard): move the address country select out of the hooks module
CountrySelect was a React component inside src/lib/hooks/, which forced
use-available-countries to become a .tsx file. Move it to
components/shared/address-country-select.tsx and give the hook module its
.ts extension back.

Named AddressCountrySelect rather than CountrySelector, because
components/shared/country-selector.tsx already exports CountrySelector -
the searchable picker for adding countries to a zone.

Render a Skeleton while the countries load, as ZoneSelector does, so a
saved country no longer shows the placeholder for a beat before its name
arrives. With the countries always present at render, SelectValue resolves
the label from the items map, so the render prop that duplicated that
lookup is gone.

Give the trigger an accessible name. The visible label from
FormFieldWrapper points at an id that renderFormControl={false} never
applies, so the field had no accessible name at all.

Drop use-available-countries.spec.ts. It mocked api.query, then asserted
that the function returned what the mock was told to return; it covered
neither the sort nor the filter nor the query key. fetchAvailableCountries
existed only to make that test possible, so it is inlined back into the
queryFn.

Relates to #5191
